--- downloader.py
@@ -1,18 +1,0 @@
-from twython import Twython
-from pprint import pprint
-from config import twitter_credentials
-import urllib.request
-
-
-twitter = Twython(
-    twitter_credentials.TWITTER_CONSUMER_KEY, twitter_credentials.TWITTER_CONSUMER_SECRET,
-    twitter_credentials.TWITTER_ACCESS_KEY, twitter_credentials.TWITTER_ACCESS_SECRET)
-
-def download(id, filename):
-    tweet = twitter.show_status(id=id)
-    pprint(tweet['text'])
-
-    # problem: twitter doesn't allow you to fetch the raw video for some tweets
-    video_url = tweet['extended_entities']['media'][0]['video_info']['variants'][0]['url']
-
-    urllib.request.urlretrieve(video_url, filename)
